requirements/base.txt
Django==1.6.5
bpython==0.13
django-braces==1.4.0
django-model-utils==2.0.3
logutils==0.3.3
South==0.8.4
psycopg2
djangorestframework<3.0